More modular build process
