+++ client.c
@@ -10,13 +10,12 @@
 #include <unistd.h>
 #include "lib/request.h"
 #include "lib/response.h"
+#include "lib/networklib.h"
 
 #include <sys/ioctl.h>
 
 const char *HOST = "localhost";
 const char *PORT = "9432";
-
-FILE *socket_dial(const char *host, const char *port);
 
 int main(int argc, char *argv[]) {
 
@@ -60,53 +59,5 @@ int main(int argc, char *argv[]) {
     return EXIT_SUCCESS;
 }
 
-FILE *socket_dial(const char *host, const char *port) {
-    /* Lookup server address information */
-    struct addrinfo *results;
-    struct addrinfo  hints = {
-        .ai_family   = AF_UNSPEC,   /* Return IPv4 and IPv6 choices */
-        .ai_socktype = SOCK_STREAM, /* Use TCP */
-    };
-    int status;
-    if ((status = getaddrinfo(host, port, &hints, &results)) != 0) {
-        fprintf(stderr, "getaddrinfo failed: %s\n", gai_strerror(status));
-        return NULL;
-    }
-
-    /* For each server entry, allocate socket and try to connect */
-    int client_fd = -1;
-    for (struct addrinfo *p = results; p != NULL && client_fd < 0; p = p->ai_next) {
-      /* Allocate socket */
-      if ((client_fd = socket(p->ai_family, p->ai_socktype, p->ai_protocol)) < 0) {
-          fprintf(stderr, "Unable to make socket: %s\n", strerror(errno));
-          continue;
-      }
-
-      /* Connect to host */
-      if (connect(client_fd, p->ai_addr, p->ai_addrlen) < 0) {
-          fprintf(stderr, "Unable to connect to %s:%s: %s\n", host, port, strerror(errno));
-          close(client_fd);
-          client_fd = -1;
-          continue;
-      }
-    }
-
-    /* Release allocate address information */
-    freeaddrinfo(results);
-
-    if (client_fd < 0) {
-    	return NULL;
-    }
-
-    /* Open file stream from socket file descriptor */
-    FILE *client_file = fdopen(client_fd, "w+");
-    if (client_file == NULL) {
-        fprintf(stderr, "Unable to fdopen: %s\n", strerror(errno));
-        close(client_fd);
-        return NULL;
-    }
-
-    return client_file;
-}
 
 /* vim: set expandtab sts=4 sw=4 ts=8 ft=c: */
