+++ lib/networklib.h
@@ -1,0 +1,4 @@
+FILE * socket_dial(const char * host, const char * port);
+FILE * accept_client(int server_fd);
+int socket_listen(const char * host, const char * port);
+FILE *socket_dial(const char *host, const char *port);
